+++ src/main/java/act/e2e/RequestSpec.java
@@ -30,7 +30,10 @@ import org.osgl.http.H;
 import org.osgl.util.C;
 import org.osgl.util.E;
 
-import java.util.*;
+import java.util.ArrayList;
+import java.util.LinkedHashMap;
+import java.util.List;
+import java.util.Map;
 
 public class RequestSpec implements InteractionPart {